rename koExtension to ko-extension
